Updated register mysql table/connection
My god it's finally over. I spent the enite day getting things running for the website again and now I just want to roll over and die. Everything is functioning again like it was and is now hopefully much more stable now that the website is being hosted on linux rather than windows with xampp. I added the loggedin column to our table as well as updated the password used to access our database.