Finished NameGen() + New Functions GenHelper() and Delete()
The NameGen() function is finally finished supporting making unique file names for footage, images, and clips. The function GenHelper() was made to not repeat code needed to get a list of the footage directory. Delete() was introduced to deal with removing old footage. The function should be operable with deleting images and clips as well, but needs testing.