Very Buggy but Functional Clips!
Clips are now recorded until motion stops and then uploaded to the server just like images! Some other changes were made such as now putting images in the image directory. While this is nice, there is still a ton to fix about clips and recording video in general such as choosing the correct framerate, testing on the raspberry pi, and getting better motion detection. There is also more features planned for clipping that have yet to be implemented such as adding on pre-motion detection segments and recording until motion has ended for a few seconds.