Added Conversion to Webm Video
Because the website has trouble diaplaying anything other than webm video and opencv is unable to record proper webm, videos clips are converted using a server side script 'called converer.py'. This should 'hopefully' be a temporary fix and well be able to get opencv to record webm or display other video types on the webstie.