Fixed Registration Bug
Turns out the error that later appeared when registering was browsers not liking that ssh was being used in the registration script to make a user directory. The issue was solved by making a seperate php script to do this and calling it via terminal from the first one. Why this wan't a problem from the start I don't know.